Consolidate styles & layout
