+++ components/layout.js
@@ -1,5 +1,8 @@
 import Head from 'next/head';
-import styles from '../styles/components/layout.module.scss';
+import Image from 'next/image'
+import bannerPic from '../public/bcc_banner.svg'
+
+import styles from '../styles/styles.module.scss'
 
 export default function Layout({ children }) {
   return (
@@ -8,6 +11,15 @@ export default function Layout({ children }) {
         <html lang="eng-US" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
+      <header className={styles.banner_container}>
+        <Image
+          src={bannerPic}
+          layout='fill'
+          objectFit='contain'
+          alt=""
+        />
+
+      </header>
       <main>
         {children}
       </main>
