+++ pages/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import Layout from '../components/layout'
 import Image from 'next/image'
-import bannerPic from '../public/bcc_banner.png'
+import bannerPic from '../public/bcc_banner.svg'
 
 import styles from '../styles/styles.module.scss'
