+++ styles/components/0_components.scss
@@ -1,0 +1,2 @@
+@import './banner.scss';
+@import './layout.module.scss';
