+++ styles/components/banner.scss
@@ -1,7 +1,3 @@
 .banner_container {
-  width: 100%;
-
-  img {
-    width: 100%;
-  }
+  width: 50%;
 }
